Make the withDocs argument default to true
Quite a few people seem to be using this nix expression and just try to build GHC without creating an `mk/build.mk`. In this configuration, the make build system goes ahead and tries to build haddocks, the sphinx manual and what not. It's therefore rather sensible to support this out of the box without having to tweak any argument.