Build and test GHC in CI
To keep the CI stable, only run a meaningful subset of tests. (We don't
want to show that GHC has no bugs. Instead, we want to show that the
ghx.nix env can build a working GHC.)

This slows down the CI. However, as there are at most couple of PRs per
month, this should be affordable. Giving maintainers more confidence on
reviewing and merging PRs should be worth it.