Fix build failures on `aarch64-darwin`
The main change here is enabling the separate `bin`
output for `ormolu`, which is currently necessary to get
HLS working on `aarch64-darwin`.  See:

https://github.com/NixOS/nixpkgs/issues/140774#issuecomment-1365283244

Adding in the overlay to fix that entailed some surgery to
`ghc.nix` and `flake.nix`, but I tried my best to make it as
unintrusive as possible.